Min-Project/requirements.txt
flask==3.0.0
groq==0.13.0
pyttsx3==2.90
markdown==3.5.2
textblob==0.17.1
python-dotenv==1.0.0 